+++ plugin/src/mosaic/plugins/LinearMovementFPTracker_3D.java
@@ -257,7 +257,7 @@ public class LinearMovementFPTracker_3D extends PFTracking3D {
 			for(float[] vDataPoint : vImageThresholded){
 				int vCentroidIndex = 0;
 				int vBestIndex = -1;
-				float vZScale = (float)(getPixelDepthInNm() / getPixelWidthInNm());
+				float vZScale = (getPixelDepthInNm() / getPixelWidthInNm());
 				float vBestScore = mWidth*mWidth + mHeight*mHeight + mNSlices * mNSlices * vZScale * vZScale;
 				for(float[] vCentroid : vCentroids){
 					float vScore = (vCentroid[0]-vDataPoint[0]) * (vCentroid[0]-vDataPoint[0]) +
